fix cleanup of pre-existing 'centos' user

+++ scripts/cleanup.sh
@@ -1,11 +1,14 @@
-#!/bin/sh -eux
+#!/bin/bash -eux
 
 rm -f /etc/udev/rules.d/70-persistent-net.rules;
 yum -y clean all
 
 # cleanup centos user account (if it exists)
 if id -u centos > /dev/null 2>&1; then
-    /usr/sbin/userdel -r centos
+    # and we are not currently logged in as the centos user...
+    if [[ $(id -u -n) == 'centos' ]]; then
+        /usr/sbin/userdel -r centos
+    fi
 fi
 
 # cleanup any ssh keys nova/cloud-init may have injected
